requirements.txt
# Core ML libraries
tensorflow>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data handling
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For data download
requests>=2.31.0